% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDirectory.r
\docType{class}
\name{AlgorithmiaDirectoryIterator-class}
\alias{AlgorithmiaDirectoryIterator}
\alias{AlgorithmiaDirectoryIterator-class}
\title{DirectoryIterator object to iterate over child files or directories of a parent directory.
To create one, call: `dataDirectory$dirs()` or `dataDirectory$files()`}
\description{
DirectoryIterator object to iterate over child files or directories of a parent directory.
To create one, call: `dataDirectory$dirs()` or `dataDirectory$files()`
}
\section{Fields}{

\describe{
\item{\code{dataDirectory}}{Parent directory whose children we are iterating over.}

\item{\code{typeFilter}}{Either DATA_OBJECT_TYPE_FILE or DATA_OBJECT_TYPE_DIRECTORY.}

\item{\code{marker}}{Page marker while iterating over a directory with lots of entries.}

\item{\code{first}}{Whether we have asked the server for anything for the first time.}

\item{\code{index}}{The current position of the directoryEntries list.}

\item{\code{directoryEntries}}{The list of children entries we are iterating over.}
}}
\section{Methods}{

\describe{
\item{\code{getNext()}}{Returns the next element or stops if you have asked for a next element that does not exist.}

\item{\code{hasNext()}}{Returns TRUE if there are any more elements. FALSE otherwise.}
}}

