% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Acl.r
\docType{data}
\name{ReadAcl.PRIVATE}
\alias{ReadAcl.PRIVATE}
\title{The ACL that allows only you to read an item.
Sample usage:
dataDirectory$create(ReadAcl.PRIVATE)
dataDirectory$updatePermissions(ReadAcl.PRIVATE)}
\format{
An object of class \code{AlgorithmiaAcl} of length 1.
}
\usage{
ReadAcl.PRIVATE
}
\description{
The ACL that allows only you to read an item.
Sample usage:
dataDirectory$create(ReadAcl.PRIVATE)
dataDirectory$updatePermissions(ReadAcl.PRIVATE)
}
\keyword{datasets}
