% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Client.r
\name{getAlgorithmiaClient}
\alias{getAlgorithmiaClient}
\title{Creates a new Algorithmia Client which you can use to call algorithms and
interact with directories and files in the Algorithmia data API.}
\usage{
getAlgorithmiaClient(apiKey = NA_character_, apiAddress = NA_character_)
}
\arguments{
\item{apiKey}{The Algorithmia API key. You need to set this when you are
interacting with Algorithmia outside of an algorithm. To find your
Algorithmia API key visit: https://algorithmia.com/users/[YOUR USER NAME]}

\item{apiAddress}{The Algorithmia API address. Normal users should not set
this. This defaults to "https://api.algorithmia.com" when it is not
explicitly set.}
}
\value{
A new AlgorithmiaClient object
}
\description{
Creates a new Algorithmia Client which you can use to call algorithms and
interact with directories and files in the Algorithmia data API.
}
\examples{
client <- algorithmia::getAlgorithmiaClient() # Inside an Algorithmia algorithm
client <- algorithmia::getAlgorithmiaClient("YOUR_ALGORITHMIA_API_KEY") # Everywhere else
}
